Decompose home view
