+++ Sources/App/HomeView.swift
@@ -525,204 +525,6 @@ extension HomeView {
         }
     }
 
-    static func selectProjection(
-        from projections: [HomeProjectionRecord],
-        currentContext: LocationContext?
-    ) -> HomeProjectionRecord? {
-        if let currentContext {
-            let projectionKey = HomeProjection.projectionKey(for: currentContext)
-            return projections.first(where: { $0.projectionKey == projectionKey })
-        }
-
-        return projections.max(by: { $0.updatedAt < $1.updatedAt })
-    }
-
-    static func selectStormSetup(
-        projection: HomeProjectionRecord?,
-        currentContext: LocationContext?,
-        pipelineValue: StormSetupDTO?,
-        pipelineRefreshKey: LocationContext.RefreshKey?,
-        now: Date
-    ) -> StormSetupDTO? {
-        if let currentContext {
-            let currentRefreshKey = currentContext.refreshKey
-            if pipelineRefreshKey == currentRefreshKey,
-               let pipelineValue,
-               pipelineValue.freshness.expiresAt > now,
-               pipelineValue.h3Cell == currentContext.h3Cell {
-                return pipelineValue
-            }
-
-            guard let projection,
-                  projection.projectionKey == HomeProjection.projectionKey(for: currentContext),
-                  let stormSetup = projection.stormSetup,
-                  stormSetup.freshness.expiresAt > now,
-                  stormSetup.h3Cell == currentContext.h3Cell else {
-                return nil
-            }
-
-            return stormSetup
-        }
-
-        guard let stormSetup = projection?.stormSetup,
-              stormSetup.freshness.expiresAt > now else {
-            return nil
-        }
-
-        return stormSetup
-    }
-
-    static func selectStormSetupCurrentResponse(
-        projection: HomeProjectionRecord?,
-        currentContext: LocationContext?,
-        pipelineValue: StormSetupCurrentResponse?,
-        pipelineRefreshKey: LocationContext.RefreshKey?,
-        now: Date
-    ) -> StormSetupCurrentResponse? {
-        if let currentContext {
-            let currentRefreshKey = currentContext.refreshKey
-            if pipelineRefreshKey == currentRefreshKey,
-               let pipelineValue,
-               pipelineValue.setup.freshness.expiresAt > now,
-               pipelineValue.setup.h3Cell == currentContext.h3Cell {
-                return pipelineValue
-            }
-
-            guard let projection,
-                  projection.projectionKey == HomeProjection.projectionKey(for: currentContext),
-                  let response = projection.stormSetupCurrentResponse,
-                  response.setup.freshness.expiresAt > now,
-                  response.setup.h3Cell == currentContext.h3Cell else {
-                return nil
-            }
-            return response
-        }
-
-        guard let response = projection?.stormSetupCurrentResponse,
-              response.setup.freshness.expiresAt > now else {
-            return nil
-        }
-        return response
-    }
-
-    static func resolveLocationTimeZone(
-        selectedProjection: HomeProjectionRecord?,
-        currentContext: LocationContext?,
-        newestStartupProjection: HomeProjectionRecord?,
-        fallback: TimeZone = .autoupdatingCurrent
-    ) -> TimeZone {
-        if let timeZoneIdentifier = selectedProjection?.timeZoneId,
-           let timeZone = TimeZone(identifier: timeZoneIdentifier) {
-            return timeZone
-        }
-
-        if let currentContext,
-           let timeZoneIdentifier = currentContext.grid.timeZoneId,
-           let timeZone = TimeZone(identifier: timeZoneIdentifier) {
-            return timeZone
-        }
-
-        if currentContext == nil,
-           let timeZoneIdentifier = newestStartupProjection?.timeZoneId,
-           let timeZone = TimeZone(identifier: timeZoneIdentifier) {
-            return timeZone
-        }
-
-        return fallback
-    }
-
-    static func selectProjection(
-        from projections: [HomeProjection],
-        currentContext: LocationContext?
-    ) -> HomeProjection? {
-        if let currentContext {
-            let projectionKey = HomeProjection.projectionKey(for: currentContext)
-            return projections.first(where: { $0.projectionKey == projectionKey })
-        }
-
-        return projections.max(by: { $0.updatedAt < $1.updatedAt })
-    }
-
-    static func showsBootstrapLoading(
-        readinessState: SummaryReadinessState,
-        isRefreshInFlight: Bool,
-        hasProjection: Bool
-    ) -> Bool {
-        readinessState != .locationUnavailable &&
-        hasProjection == false &&
-        (isRefreshInFlight || readinessState != .ready)
-    }
-
-    static func preferredSummaryValue<T>(
-        projectionValue: T?,
-        pipelineValue: T?,
-        prefersPipelineValue: Bool
-    ) -> T? {
-        if prefersPipelineValue {
-            return pipelineValue ?? projectionValue
-        }
-        return projectionValue ?? pipelineValue
-    }
-
-    static func shouldRefreshStormSetupSettings(
-        previousPreferences: StormSetupPreferences?,
-        currentPreferences: StormSetupPreferences
-    ) -> Bool {
-        previousPreferences != currentPreferences
-    }
-
-    static func preferredOutlooks(
-        cachedOutlooks: [ConvectiveOutlookDTO],
-        liveOutlooks: [ConvectiveOutlookDTO]
-    ) -> [ConvectiveOutlookDTO] {
-        liveOutlooks.isEmpty ? cachedOutlooks : liveOutlooks
-    }
-
-    static func preferredOutlook(
-        cachedOutlook: ConvectiveOutlookDTO?,
-        liveOutlooks: [ConvectiveOutlookDTO],
-        liveOutlook: ConvectiveOutlookDTO?
-    ) -> ConvectiveOutlookDTO? {
-        liveOutlooks.first ?? cachedOutlook ?? liveOutlook
-    }
-
-    struct LocationReliabilityRailState: Equatable {
-        let shouldShowRail: Bool
-        let qualifyingDay: String?
-        let shouldRecordImpression: Bool
-    }
-
-    static func locationReliabilityRailState(
-        reliability: LocationReliabilityState,
-        stormRisk: StormRiskLevel?,
-        severeRisk: SevereWeatherThreat?,
-        ledger: LocationReliabilityAskLedgerSnapshot,
-        now: Date,
-        timeZone: TimeZone,
-        currentlyShownQualifyingDay: String?
-    ) -> LocationReliabilityRailState {
-        let decision = LocationReliabilitySummaryRailEligibility.decision(
-            reliability: reliability,
-            stormRisk: stormRisk,
-            severeRisk: severeRisk,
-            ledger: ledger,
-            now: now,
-            timeZone: timeZone
-        )
-
-        guard decision.isEligible else {
-            return .init(shouldShowRail: false, qualifyingDay: nil, shouldRecordImpression: false)
-        }
-
-        let qualifyingDay = LocationReliabilitySummaryRailEligibility.localDayString(for: now, timeZone: timeZone)
-        let shouldRecordImpression = currentlyShownQualifyingDay != qualifyingDay
-        return .init(
-            shouldShowRail: true,
-            qualifyingDay: qualifyingDay,
-            shouldRecordImpression: shouldRecordImpression
-        )
-    }
-
     private func refreshLocationReliabilityRail() {
         if isUITestForceReliabilityRail {
             let qualifyingDay = LocationReliabilitySummaryRailEligibility.localDayString(
@@ -823,201 +625,4 @@ extension HomeView {
         LocationReliabilityAskLedger.live().recordSameDaySuppression(qualifyingDay: qualifyingDay)
     }
 
-}
-
-private struct TodayTabView: View {
-    @State private var headerCondenseProgress: CGFloat = 0
-    @State private var visibleWeatherState = TodayVisibleWeatherState()
-
-    let snap: LocationSnapshot?
-    let stormSetup: StormSetupDTO?
-    let stormSetupProfileAnalysisResponse: AnvilAnalyzeProfileResponse?
-    let stormSetupPreferences: StormSetupPreferences
-    let stormRisk: StormRiskLevel?
-    let severeRisk: SevereWeatherThreat?
-    let fireRisk: FireRiskLevel?
-    let mesos: [MdDTO]
-    let alerts: [AlertDTO]
-    let outlook: ConvectiveOutlookDTO?
-    let weather: SummaryWeather?
-    let airQuality: AirQualityCurrentResponse?
-    let locationTimeZone: TimeZone
-    let todayContentState: TodayContentState
-    let localAlertsDisplayState: LocalAlertsDisplayState
-    let readinessState: SummaryReadinessState
-    let resolutionState: SummaryResolutionState
-    let isRefreshInFlight: Bool
-    let showsOfflineToken: Bool
-    let locationReliabilityRailState: SummaryView.LocationReliabilityRailState?
-    let onOpenMapLayer: (MapLayer) -> Void
-    let onOpenAlerts: () -> Void
-    let onOpenOutlooks: () -> Void
-    let refreshAction: () async -> Void
-
-    private var weatherLocationIdentity: SummaryWeatherLocationIdentity? {
-        SummaryWeatherLocationIdentity(snapshot: snap)
-    }
-
-    private var visibleWeather: SummaryWeather? {
-        TodayVisibleWeatherState.resolve(
-            liveWeather: weather,
-            displayedWeather: visibleWeatherState.weather,
-            isRefreshing: isRefreshInFlight,
-            displayedWeatherLocationIdentity: visibleWeatherState.locationIdentity,
-            weatherLocationIdentity: weatherLocationIdentity
-        ).weather
-    }
-
-    private var visibleWeatherTaskState: TodayVisibleWeatherStateTaskState {
-        TodayVisibleWeatherStateTaskState(
-            liveWeather: weather,
-            isRefreshing: isRefreshInFlight,
-            weatherLocationIdentity: weatherLocationIdentity
-        )
-    }
-
-    var body: some View {
-        NavigationStack {
-            ScrollView {
-                SummaryView(
-                    snap: snap,
-                    stormSetup: stormSetup,
-                    stormSetupProfileAnalysisResponse: stormSetupProfileAnalysisResponse,
-                    stormSetupPreferences: stormSetupPreferences,
-                    stormRisk: stormRisk,
-                    severeRisk: severeRisk,
-                    fireRisk: fireRisk,
-                    mesos: mesos,
-                    alerts: alerts,
-                    outlook: outlook,
-                    weather: visibleWeather,
-                    airQuality: airQuality,
-                    locationTimeZone: locationTimeZone,
-                    todayContentState: todayContentState,
-                    localAlertsDisplayState: localAlertsDisplayState,
-                    readinessState: readinessState,
-                    resolutionState: resolutionState,
-                    showsOfflineToken: showsOfflineToken,
-                    headerCondenseProgress: headerCondenseProgress,
-                    locationReliabilityRailState: locationReliabilityRailState,
-                    onOpenMapLayer: onOpenMapLayer,
-                    onOpenAlerts: onOpenAlerts,
-                    onOpenOutlooks: onOpenOutlooks
-                )
-                    .toolbar(.hidden, for: .navigationBar)
-                    .background(.skyAwareBackground)
-            }
-            .accessibilityIdentifier("summary-scroll")
-            .onScrollGeometryChange(for: CGFloat.self) { geometry in
-                geometry.contentOffset.y + geometry.contentInsets.top
-            } action: { _, newValue in
-                let normalizedProgress = min(max((newValue - 6) / 68, 0), 1)
-                if abs(headerCondenseProgress - normalizedProgress) > 0.001 {
-                    headerCondenseProgress = normalizedProgress
-                }
-            }
-            .background(Color(.skyAwareBackground).ignoresSafeArea())
-            .refreshable {
-                await refreshAction()
-            }
-        }
-        .background(Color(.skyAwareBackground).ignoresSafeArea())
-        .task(id: visibleWeatherTaskState) {
-            visibleWeatherState = TodayVisibleWeatherState.resolve(
-                liveWeather: weather,
-                displayedWeather: visibleWeatherState.weather,
-                isRefreshing: isRefreshInFlight,
-                displayedWeatherLocationIdentity: visibleWeatherState.locationIdentity,
-                weatherLocationIdentity: weatherLocationIdentity
-            )
-        }
-    }
-}
-
-private struct TodayVisibleWeatherStateTaskState: Equatable {
-    let liveWeather: SummaryWeather?
-    let isRefreshing: Bool
-    let weatherLocationIdentity: SummaryWeatherLocationIdentity?
-}
-
-#Preview("Home") {
-    HomeView(
-        initialSnap: .init(
-            coordinates: .init(latitude: 39.75, longitude: -104.44),
-            timestamp: .now,
-            accuracy: 20,
-            placemarkSummary: "Bennett, CO"
-        ),
-        initialStormRisk: .slight,
-        initialSevereRisk: .tornado(probability: 0.10),
-        initialFireRisk: .extreme,
-        initialMesos: MD.sampleDiscussionDTOs,
-        initialAlerts: Watch.sampleWatchRows,
-        initialOutlooks: ConvectiveOutlook.sampleOutlookDtos,
-        initialOutlook: ConvectiveOutlook.sampleOutlookDtos.first
-    )
-    .environment(\.dependencies, Dependencies.unconfigured)
-    .environment(LocationSession.preview)
-    .environment(RemoteAlertPresentationState())
-    .environment(RuntimeConnectivityState.preview)
-    .modelContainer(HomeViewPreviewData.modelContainer)
-}
-
-@MainActor
-private enum HomeViewPreviewData {
-    static let modelContainer: ModelContainer = {
-        let schema = Schema([HomeProjection.self, ConvectiveOutlook.self])
-        let configuration = ModelConfiguration(isStoredInMemoryOnly: true)
-        let container = try! ModelContainer(for: schema, configurations: configuration)
-        let context = ModelContext(container)
-
-        if let previewContext = LocationSession.preview.currentContext {
-            let projection = HomeProjection(
-                context: previewContext,
-                createdAt: .now,
-                lastViewedAt: .now
-            )
-            projection.weatherPayload = HomeProjectionWeatherPayload(
-                summary: SummaryWeather(
-                    temperature: .init(value: 72, unit: .fahrenheit),
-                    symbolName: "sun.max.fill",
-                    conditionText: "Clear",
-                    asOf: .now,
-                    dewPoint: .init(value: 54, unit: .fahrenheit),
-                    humidity: 0.45,
-                    windSpeed: .init(value: 15, unit: .milesPerHour),
-                    windGust: .init(value: 24, unit: .milesPerHour),
-                    windDirection: "NW",
-                    pressure: .init(value: 29.92, unit: .inchesOfMercury),
-                    pressureTrend: "steady"
-                )
-            )
-            projection.stormRisk = .slight
-            projection.severeRisk = .tornado(probability: 0.10)
-            projection.fireRisk = .extreme
-            projection.activeMesos = MD.sampleDiscussionDTOs
-            projection.activeAlerts = Watch.sampleWatchRows
-            projection.updatedAt = .now
-            context.insert(projection)
-        }
-
-        if let outlook = ConvectiveOutlook.sampleOutlookDtos.first {
-            context.insert(
-                ConvectiveOutlook(
-                    title: outlook.title,
-                    link: outlook.link,
-                    published: outlook.published,
-                    fullText: outlook.fullText,
-                    summary: outlook.summary,
-                    day: outlook.day,
-                    riskLevel: outlook.riskLevel,
-                    issued: outlook.issued ?? outlook.published,
-                    validUntil: outlook.validUntil ?? outlook.published
-                )
-            )
-        }
-
-        try! context.save()
-        return container
-    }()
 }
